Refactor: 콘텐트 이름 변경
